requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
bcrypt==4.1.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==41.0.7
ecdsa==0.19.1
fastapi==0.104.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
passlib==1.7.4
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
python-jose[cryptography]==3.3.0
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
twilio==8.10.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
yarl==1.20.1
pydantic-settings
